Re-use same figure table
