--- src/components/tables/EntriesTable/EntriesFigure/styles.css
@@ -1,14 +1,0 @@
-.container {
-    flex:1;
-}
-
-.content {
-    display: flex;
-    position: relative;
-    flex-direction: column;
-
-    .table {
-        width: 100%;
-        overflow-x: auto;
-    }
-}
