+++ src/components/tables/EntryFiguresTable/styles.css
